Add files via upload
Sentiment analysis is a technique that detects the underlying sentiment in a piece of text.it is the process of classifying text as either positive, negative, or neutral. Machine learning techniques are used to evaluate a piece of text and determine the sentiment behind it whether it is positive or neutral or negative.
I collected the data from twitter and analysis that according to how many positive neutral, and negative tweets are tweeted. I show the difference between the accuracy score with different ML algorithms.